+++ src/maths/v3_init.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   v3_init.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: flfische <flfische@student.42heilbronn.    +#+  +:+       +#+        */
+/*   By: klamprak <klamprak@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/11 09:58:13 by flfische          #+#    #+#             */
-/*   Updated: 2024/06/11 09:59:59 by flfische         ###   ########.fr       */
+/*   Updated: 2024/06/27 16:36:43 by klamprak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
  * @param z The z coordinate.
  * @return The 3D vector.
  */
-t_vector3	*ft_v3_init(t_vector3 *v, float x, float y, float z)
+t_vector3	*ft_v3_init(t_vector3 *v, double x, double y, double z)
 {
 	v->x = x;
 	v->y = y;
